Updated openstack/openstack Project: openstack/cinder  b7d778d5c4dccf364d4328c07a6b0cdf5e5f24cd
Log message cleanup for volume-usage-audit

volume-usage-audit utlis logs unecessary message at for deleted
volumes or exception:

  try:
      LOG.debug('Send notification for vol X')
      _notify_func
  except Exception:
      LOG.exception(...)

In the case when _notify_func raises exception, it logs
'Send notification' and the exception.

This change reduce logs by only logging 'Sent notification' when
there is no exception by:

  try:
      _notify_func
      LOG.debug('Sent notification for vol X')
  except Exception:
      LOG.exception(...)

Also, this change fixes notify_snapshot_usage() for not handling
the case when source volume for the snapshot is deleted.

Closes-bug: #1631561

Change-Id: Iff5448e888ee39bce1ff99c7bf48581d655fb687