Updated openstack/openstack Project: openstack/nova-specs  339c9a5e076e971ab7619d3d995cdcbbdfc0a21a
Amend placement API with correct PUT inventory

At the end of the review for the generic resource pools spec, we agreed
to change the JSON structure of the inventory "records" to this:

 {
    "DISK_GB": {
        "total": 0,
        ...
    }, ...
 }

instead of this:

 [
    {
        "resource_class": "DISK_GB",
        "total": 0,
        ...
    }, ...
 ]

However, we forgot to make the PUT /r-p/{uuid}/inventories request
consistent with the GET response payload.

This corrects that oversight. The code in the placement API was written
to the spec and will need to be corrected as well.

Change-Id: I8eaef8d532d9e5d4917ede572000e3adc3149f73