Updated openstack/openstack Project: openstack/cinder  2abff84fa076c43c604ef4c7a347669065c913d1
Pass OVO instance to copy_volume_to_image

We should be using OVO instances as much as possible in the volume
manager's code, but we still have some instances of ORM instances being
passed to drivers.

This patch fixes only 1 specific case, so there are more cases still
pending to change to OVOs, which is the call from the
copy_volume_to_image method in the manager to the respective driver
method.

Closes-Bug: #1641611
Change-Id: If30d3521d3a28488b12ed3bd4c753066b80d39c9