Updated openstack/openstack Project: openstack/nova  742414bf7c6c2a786538cc8b2cf0397e27bad0e2
Transform requested secgroup names to uuids

We currently don't store requested security groups with the instance
(or build request) when using neutron, but we plan on doing that as
part of the cells v2 scheduling interaction and moving the instance
create to conductor. The plan is to store the SecurityGroupList
in the request spec so we can lookup the requested security groups
again later.

In order to avoid a situation where the name changes on a security
group in neutron, we need to store the uuid which is immutable. So
this change transforms any requested security group names to uuids
when using neutron as if the user had only requested uuids.

Nothing is using this yet as populate_security_groups in the
neutron security group driver is basically a no-op, but later
changes will rely on this when storing the security groups in the
request spec.

Part of blueprint cells-scheduling-interaction

Change-Id: I1964ee3a7529bb3a1facd38ee62a71c148ecaef1