Updated openstack/openstack Project: openstack/diskimage-builder  be1e5635245e57e5bea721b03ba153676e7ceff7
simplify ARCH param for rhel/centos param can be x86_64 and amd64

for fedora/rhel/centos the main supported ARCH is x86_64. This patch allow
to call diskimage-builder with the above distro's with param ARCH=x86_64,
And also retain same behaiver when call with ARCH=amd64 as it translate
anyway to x86_64. Doing so wil simplify user expirience.

Change-Id: I229e0912434109b1b48a030bd35ad8dc1096a629