Updated openstack/openstack Project: openstack/heat  013207210640ef8ceb9fd5db46757d10fd30ebc7
Don't delete RawTemplate if it referenced by a stack

When creating a nested stack and pre-storing the raw template, we
previously used a try/finally block to ensure that if the stack creation
was never started (e.g. because the template failed validation at the other
end) that we would delete the raw template again rather than leave it lying
around in the DB. However, this catches too broad a range of exceptions. If
an exception such as GreenletExit occurs inside this block, then we will
attempt to delete the raw template, resulting in a DB IntegrityError either
here (because the template id is already referenced by a stack) or when
trying to store the stack (because the template is gone).

Resolve this by only doing the cleanup on HeatExceptions, which indicates
that the other engine has bailed out without storing the stack.

Change-Id: Ie24a82b7ce16281c72d4eb66cca8683248628b46
Closes-Bug: #1626256