Updated openstack/openstack Project: openstack-infra/project-config  1288b8d704932874ee79c3c78e7e23b165041acd
Introduce TripleO scenario004

This is meant to run newer services that are not available in
releases older than Ocata such as Barbican.

So, for instance, Barbican was moved from scenario002 (where it no
longer runs) to scenario004). Also, move Mistral and Zaqar, as these
either haven't landed or landed in Ocata.

Change-Id: I929a2345e817b2b198d26fafaa283a6fbffe3fe9
Depends-On: I19ed2dddd88ae1752c6b7500c1790f801f6bfcd9