Updated openstack/openstack Project: openstack-infra/project-config  e64580eb7349158b155955695a8245f3bc13ac83
install distro packages before running the announce script

The swift client release [1] failed because there was a missing system
dependency. We were already installing the dependencies of the thing
being released, but not of the release tools themselves. This change
adds the extra step to ensure that the dependencies of the release tools
are available before they are run.

[1] http://logs.openstack.org/8e/8e127421b7cf0c5fe19b0aae3bcc0fdd12c4c169/release/python-swiftclient-announce-release/d5aefbf/console.html

Change-Id: Iea5513fd7734e493d2e24ce3e8fb4b3887ef79ec