Updated openstack/openstack Project: openstack-infra/project-config  af47eecf26a29e8758658774d2d932aaafba0014
Add up-to-date check to oaktree-diskcheck job

The generated go files need to be committed to the repo along with
changes to the proto files. (This isn't my first choice - but because of
the way library depends work in go it's the only thing that makes sense)

Because of that, every change to a proto file should cause the developer
to run make, produce the go changes and commit them along with the proto
changes. If the go files are up to date, running make in a git checkout
should not cause files to change.

Add a quick check to make sure people don't forget the go code. (I've
already done this once)

Change-Id: I50c6589f6b713c297d8827629ea13e6b68f72228