Updated openstack/openstack Project: openstack/tripleo-puppet-elements  bf992a9e90658af32df36c374176ec327d330fef
Pin puppetlabs-ntp when installing from source

Due to the puppet3 support removal in the upstream puppetlabs-ntp
module, we need to provide a default pin for the module when a developer
is installing from source.  If the user does not properly pin the ntp
module, the overcloud deployment fails due to a puppet parse error.

Change-Id: I2f4c1387b61c4aa2dcbb1ec5a3fa298e9330f0e8
Closes-Bug: #1633713