Updated openstack/openstack Project: openstack/neutron  d41bed0ee674e26f58e5ffe51198260ff765ccf6
Parse the output of ip route more robustly

The code parsing the output of ip route assumes that the output is the
destination network, and then key/value pairs. This is not necessarily
the case, as the output may contain flags.

This change modifies the parsing of ip route's output to look
specifically for the keys of interest, and read their values
exclusively.

Change-Id: I34b6c0afb05550c970b1cacda8ec472294215403
Closes-Bug: 1640008