Updated openstack/openstack Project: openstack-dev/devstack  29d13df1a284f8f1a5973ccc826a475156820d23
Adopt openstack client for neutron commands

The neutron client is going to be deprecated during the
Ocata timeframe, so it is time to start switching to the
openstack client to invoke networking commands.

use of neutron client in neutron-legacy has been left as is.

The command for setting the router gateway is left as follow up.

Change-Id: I0a63e03d7d4a08ad6c27f2729fc298322baab397