Updated openstack/openstack Project: openstack/governance  5f09264bfef31a3a86f54b3e655eea2b8fbc6bf6
Revisit list of governed neutron projects

Initiatives that fall short of the requirements for neutron project
inclusion and that have no proactive team ready to address the
identified gaps are removed from the neutron governance.

This patch takes neutron-vpnaas, networking-l2gw and networking-onos
out of the governance list for the Ocata release.

Change-Id: I4100079659f4a30729e3c2e16c1e39c2a067a974
Depends-on: Ibe5f167d2d54415d2cdd672a22c4c52793035ebe
Depends-on: I13548815c4522da109d342d868ef57a69adbf9e9
Depends-on: I1fea9e95393ec1ec3c4109759d2abb839373e1c6