Updated openstack/openstack Project: openstack/governance  ec2379dc5e78f10d030830c12f40912be3ac09f9
Add openstack/molteniron to ironic governance

In the last ironic team meeting[0], the core team unanimously agreed
molteniron[1] belongs in the ironic governance, per our rules around
that.[2] Add it.

Molteniron is a tool similar to nodepool to reserve baremetal machines
for use by a CI that tests ironic. The CI does *not* run on these
machines; rather they are used as physical resources to test ironic with
real hardware. An ironic deployment + nodepool is not sufficient for
this, as the test system needs access to the out of band controller
(IPMI, iLO, DRAC, etc), and in an ironic deployment ironic controls
that instead.

This is in use by at least one third party CI, and others have plans to
use it.

[0] http://eavesdrop.openstack.org/meetings/ironic/2016/ironic.2016-11-14-17.00.html
[1] http://git.openstack.org/cgit/openstack/molteniron/
[2] http://docs.openstack.org/developer/ironic/dev/governance.html

Change-Id: I7d5b2b9b1fc1a471461ccb81bf55211e6414bf59