Updated openstack/openstack Project: openstack/cliff  9a8de2a68ac37a7931c03fc44e4817b2316e8212
Show entire command in error message

Currently only the first argument is shown in the error message.
In case of a multiword command, only first part is shown.
e.g When "openstack security grou list" command is run,
"security" subcommand is shown as invalid. This conveys a
wrong message to the user.

Closes-Bug: 1640031

Change-Id: I748ce46835638916279edead894e3f32c12abd51