Updated openstack/openstack Project: openstack/oslo.middleware  e71a24bc53150ef37036e9c196a368525f0526e4
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I399cfbc691637ef6e40ecb1e2c18a9227f84e243