Updated openstack/openstack Project: openstack-infra/project-config  5dd7714098289c28bc9be9214eda90d4c6389651
Disable devstack plugin for builder 'ironic-grenade'

Disable the devstack plugin for items which use the 'ironic-grenade'
builder as the grenade plugin in Ironic already enables the devstack
plugin. If the devstack plugin is enabled twice it gets run twice
which will have it attempting to create the bare-metal VMs twice and
failing.

Change-Id: I75ca21804891b91a928bdafe9146e27eae425f63