Updated openstack/openstack Project: openstack-infra/project-config  a976efcca57fa3683265d8f52666460333e79a88
rename 'tox' macro to 'run-tox'

JJB already has a tox module[1].  Defining a macro with the same name is
confusing and may cause problems when attempting to use this macro
elsewhere. Renaming it to something else should clear things up.

This is the warning from JJB:

$ jenkins-jobs test -o out ../project-config/jenkins/jobs
WARNING:jenkins_jobs.registry:You have a macro ('tox') defined for
'builder' component type that is masking an inbuilt definition

[1] http://docs.openstack.org/infra/jenkins-job-builder/builders.html#builders.tox

Change-Id: I6946229f6e94a5f4c10d6a5fc3b7049adeacacb1