Updated openstack/openstack Project: openstack/governance  7687889323e21899ad92756120e41c3b7669393d
Generate badges for teams and deliverables

Adds a plugin for sphinx for generating badges for teams and
deliverables to use in READMEs, docs sites, etc. A badge corresponds
to a tag.

This plugin generates an `svg` image containing team tags and the
deliverable tags for each repo listed. Each tag corresponds to an SVG
image. All the tags are bundled in the same SVG image whose filename
corresponds to the repo name.

Each badge points to the tag link in the governance website.

This not handled yet:

- This doesn't handle "unofficial" projects, which would be nice to
  have.  If a project tries to access `project-team.svg` tag, instead of
  returning 404, we could return a `project:unofficial` badge.

- Specialized colors for some tags. We might want to use different colors
  for some groups. Team tags could be orange, etc.

- Group/Order tags generation by tag groups

Here's a dummy repo created as an example of how this work would be
consumed (or look like):

- https://github.com/flaper87/badges-tester

Change-Id: Ic70c17b60c0107e40b78bf21dc3a68c558eee06f