Updated openstack/openstack Project: openstack/nova  0f0d7584e0ee4772c42d0e327624005d85108edb
Add the initial documentation for the placement API

This is the initial set of docs for the placement API
service introduced in the Newton release.

We still have to flesh out the API reference in detail
but this gets us started.

The deployment steps are taken from how devstack does
this today.

Change-Id: Ic2436d92a7cefaeb1ae67ed878da968444f2f18d