Updated openstack/openstack Project: openstack/neutron  25997b8af93b44a1e0079f89d954aad11d12962e
Fix ML2, base db plugin update_subnet for transactions

This refactors the DB base plugin update_subnet method into
precommit and postcommit methods so a core plugin can avoid
executing the postcommit methods while it has an open
transaction. Backwards compatibility is preserved for plugins
which do not need this.

This is important because the update_subnet method in the
base plugin performs operations that shouldn't be inside
of an uncommited transaction, such as updating ports,
generating AMQP notifications, and generating AFTER_UPDATE
callback events. Until this patch, ML2 was performing the
entire base plugin update_subnet operation in a transaction.

This is also necessary for the facade and push-notification
switch because ongoing db sessions will not be used for
core methods and notification generation.

Related-Bug: #1540844
Partially-Implements: blueprint push-notifications
Partially-Implements: blueprint enginefacade-switch
Change-Id: Ib825a9e0c0c6693a4e708ae6ce362775c70d843d