Updated openstack/openstack Project: openstack-infra/zuul  63a595bae3f3db85582ddba2d51ad4124877034f
Don't retry when using synchronize module

There is a bug (https://github.com/ansible/ansible/issues/18281) in the
ansible synchronize module that causes any retry attempt at
synchronizing to fail because the paths get munged resulting in invalid
paths. Unfortunately this also means that the error message we get is
not for the first failed sync attempt but for the last making it hard to
debug why things failed in the first place.

Address this by not attempting to retry until ansible is fixed. This way
we get accurate error messages more quickly (as we don't retry over and
over and generate a bad error message at the end).

Change-Id: I545c44b11f37576edc8768a3ed78962ff870995f