Updated openstack/openstack Project: openstack/neutron  ae3f3d8e61d040e645ae8668f76b6dc5fb8ffdd0
Avoid UnboundLocalErrors in _create_bulk_ml2

The exception handler blocks in _create_bulk_ml2 assumed
some local variables from iterations would be present.
This led to masking errors caused by failures in starting
the session or fetching the method with getattr because
the local variable reference would result in an
UnboundLocalError exception.

This patch just narrows the exception handlers to the
iterations.

TrivialFix

Change-Id: I74908817049826d1329a3e50cccc4c2c5708b621