Updated openstack/openstack Project: openstack/nova  b5bef5e80d770e93809c1c5cd6259dff5c2d49f1
Make n-net refuse to start unless using CellsV1

At the Barcenlona summit, we discussed disabling n-net everywhere as a
precursor to removal. The only exception for using n-net is as required
for CellsV1 users, which require it as the only out-of-the-box supported
networking option.

We will remove this check and n-net entirely when those users have a
migration path to neutron and cellsv2.

Depends-On: I47f72eac6a7657bf44319fa5d2df18ce62827f6f

Change-Id: I7309afa5230bcaeeeddb554d95c327d3a614ae6e