Updated openstack/openstack Project: openstack/python-troveclient  1ae94d4107dc0faf8d936d55f3bb398a709428c4
Get python-troveclient to use upper-constraints.txt

In review[1] which passed just fine in check some days ago failed in
gate today. This is because the current tox environment uses requests
v2.12.1 which appears to not work and is beyond the value in
upper-constraints.txt. As observed in [2] we need
upper-constraints.txt in python-troveclient.

But since python-troveclient is itself listed in upper-constraints.txt
you can't just stick it in tox.ini and expect good things to happen.

I'm adding a co-author to this change based on [2].

[1] https://review.openstack.org/#/c/382493/
[2] https://review.openstack.org/#/c/398339/

Co-Authored-By: jiansong <jian.song@easystack.cn>
Change-Id: I386efb2d5c147417af7ea25704405977c9b6bbcd