Updated openstack/openstack Project: openstack-infra/project-config  dfb9fc9848ff3aae75fcba72f8ef6b16779ce8aa
Add os_monasca-agent to OpenStack-Ansible

This patch adds a new pre-seeded git repository to OpenStack-Ansible
for deploying Monasca-Agent.

The os_monasca-agent repository has its own core team with the general
openstack-ansible core team included with the appropriate rights to
ensure that they can act as a back-stop in case of any issues.

Needed-By: I16a60601256e2e45a4c8ce8014aae9e4a87c7e8a
Change-Id: Ia6295e94c8c66377849021e86f1565c1600a6d9e