Updated openstack/openstack Project: openstack/neutron  24166e2981e389f0b9a11423d1a4617b9f67715d
Add sg_id in the AFTER_DELETE event of sg_rule delete

In the AFTER_DELETE of sg_rule delete, the sg_rule is actually deleted.
This makes it impossible/hard to know which sg is affected.

This patch add the sg_id in the event. And this patch will be used in [1].

[1] https://review.openstack.org/#/c/367718

Change-Id: Ibdef6a703913b74504e402d225b1a0dffadb7aff
Closes-Bug: #1621698