Updated openstack/openstack Project: openstack/nova-specs  148f1d76a6908170e3c9a1e6e64b46bd96345fc8
Libvirt driver emulator threads placement policy

The Nova schedular determines CPU resource utilization and instance
CPU placement based on the number of vCPUs in the flavour. A number
of hypervisors have work that is performed in the host OS on behalf
of a guest instance, which does not take place in association with
a vCPU. This is currently unaccounted for in Nova scheduling and
cannot have any placement policy controls applied.

Blueprint: libvirt-emulator-threads-policy
Change-Id: I069cd14ea89045136ae8a29c2d2f6c8c17157533