Updated openstack/openstack Project: openstack/tooz  73d76cdf2327335aeb9a16fed1e85617e172140d
Use the internal group of list rather than listing the groups

For an unknown reason, the common run_watchers implementation prefers to
list the group using the driver rather than using the groups that
actually have hooks. This simplify the code and save us from a
round-trip of getting the list of group.

Change-Id: Ice06b8282a8c48fc4963b451a8a3e03b6684e90f