Updated openstack/openstack Project: openstack/ironic  3899f9973065d9bf988e20b16ea3e7e148a0d469
Enable PXE for systems using petitboot

Petitboot systems do not use the pxelinux.0 loader. Instead,
petitboot itself handles pxelinux functionality. A side effect of
this is that petitboot requires DHCP option 210 (tftp path-prefix)
as this information is normally derived from the boot file name.
This patch uses the value of tftp_root as a path prefix and makes
the PXE module honor that prefix.

Change-Id: Ib9e954feea2cec38dd8328ada35005c0311c2a1b
Closes-Bug: #1639187
Co-Authored-By: Michael Turek <mjturek@linux.vnet.ibm.com>