Updated openstack/openstack Project: openstack/api-wg  03b5bdd43085114fa162df7f47e2263af11d5b2b
Add the operator for "not in" to the filter guideline

When filtering, one can use the ``in:`` operator to filter on results
that have a particular value in a specified list.
The question of how to specify a value that is *not* in a list came up,
so this patch adds the ``nin:`` operator to signify "not in". Note that
there was also a good deal of ambivalence between ``nin:`` and
``not_in:``, but it was felt that ``nin:`` was more consistent with
``neq:``.

Change-Id: I56049fccdea1d766fcccfd38307b02ef3baf7dd9