Updated openstack/openstack Project: openstack/nova-specs  5cd279a4688be8db3d673ee00700487bffe0ca28
Add prep-for-network-aware-scheduling-ocata spec

Change how we talk to neutron to allow us in the future to implement
network aware scheduling.

This is just the continuation of the work started during newton.

This is a priority, because this blocks routed networks, which is
blocking the removal of cells v1.

blueprint prep-for-network-aware-scheduling-ocata

Change-Id: Id28509c7fae784ecc6359556a6da1bc39f2b68f7