Updated openstack/openstack Project: openstack/diskimage-builder  f5d23c975a0ca40e720fba5afb8e5fbddba90b52
Change path for dnf arch override so basearch is not overwritten.

After writing the basearch value to /etc/dnf/vars/basearch the
arch value was overwriting the same file.  This appears to be
incorrect, so changing it to write /etc/dnf/vars/arch, which
matches the subsequent 'yum' code paths.

Change-Id: I5da54f03224c11f9e286f16b68533936c4174c2a