Updated openstack/openstack Project: openstack/diskimage-builder  0fbf13155025c8582548b295f64e67143b1e84ba
Cleanup yumdownloader repos

yumdownloader has to have all the repo XML files, etc, which adds up
to a not totally insignificant 150MiB or so.  Currently we're leaking
this directory for every build, which adds up on regualar builders
like nodepool.

Isolate the call with a separate TMPDIR so we can clean it up after
the initial download.

Change-Id: Ic65e8ca837cc76b7a1bb9f83027b4a5bdd270f75