Updated openstack/openstack Project: openstack/neutron  858e3cb47861c119d12eeacceafb40b9c440c89c
Only send string values to OVSDB other_config column

The other_config columns in OVSDB are defined as maps with string
keys and string values. The OVS agent was passing an integer
segmentation id and could pass None as the physical_network.
Unfortunately, the upstream Python OVS library does not pass the
exceptions through to us.

Change-Id: Iafa6be3749b1ee863f5fa71150c708fc46951510
Closes-Bug: #1630920