Updated openstack/openstack Project: openstack/swift  15e61c82bed8a55e57da51b26ef8486209d7dc61
Correctly send 412 Precondition Failed in copy middleware

Previously in copy middleware, if a user entered an invalid destination
path with an invalid `container/object` path the server would return
a 500 Internal Server Error. However, the correct response should be
a 412 Precondition Failed. This patch updates copy so that it catches
the 412 Precondition Failed exception and returns it to the client.

Closes-Bug: #1641980

Change-Id: Ic4677ae033d05b8730c6ad1041bd9c07268e11a9