Updated openstack/openstack Project: openstack-infra/system-config  d26f33b1ca172a70ee5e469720b4cf9d4a84c331
Set AllowOverride to FileInfo for g.o.o

In order for the `.htaccess` file to be read and processed in the g.o.o
site, we need to set AllowOverride to something else than None. FileInfo
should be enough for now as the only rule we're overriding is the
ErrorDocument[0].

This rule is needed to redirect missing badges to the "unofficial
project" one.

[0] http://git.openstack.org/cgit/openstack/governance/tree/doc/source/badges/.htaccess

Change-Id: Ie078e271732563e1a18bb0837cf9a40c6e5fcfa7