Updated openstack/openstack Project: openstack/neutron-fwaas  8bf87a0b0593d0a26604678c5c3ee170f251aa83
Set NETWORK_API_EXTENSIONS in fwaas devstack plugin

The NETWORK_API_EXTENSIONS environment variable needs to be controlled
so that the fwaas or fwaas_v2 extensions can be properly added.  This is
necessary because the tempest tests for v1 and v2 trigger based on what
extensions are loaded.  Without this, NETWORK_API_EXTENSIONS would
default to 'all', and the fwaas_v2 tests would run when fwaas v1 is
loaded and vice versa.

Change-Id: I12d765c38c1cfc7c397fef4497e9f11f260f4517
Needed-By: I8b8ddf2a9cc4d2f18c4b32917630c2a26ee0d713
Needed-By: I9fc39c5adcf136fce520c329f48cbad60cd21861