Updated openstack/openstack Project: openstack/neutron  cbbfe9b04c01142a1b1790d9fc533fc28a7d3101
Typo on side_effect rendering the test useless

side_effect does not take a trailing "s", calling
side_effects will simply ignore the call and
continue.  These test work because the no return
value are set, therefore the result is false.

Without changing the tests you can comment out
the try/except without breaking anything.

Fixing the tests shows that the call to
netifaces.ifaddresses was not covered for OSError
and test was missing for ValueError

Change-Id: I770390e99b4b614aff5512da11c0c226dc38b85f