Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  9dc5b85f64c3ddd41b9fcecc3be1b1e56d0644f0
Adds height and width support to the Plot Publisher

Height and width are supported by the Plot Plugin today but
not made available in the Publisher.  This adds support for
both with the same defaults that exist in the plugin.

This allows users to configure a height and width as desired.

Change-Id: I1846e73d2fc3f6a9f651646b74cd3af6f85aee6d