Updated openstack/openstack Project: openstack/trove  da9d3d367202ac0856caac82c3a531313279199d
Fix race condition in cluster-show

There is a race condition in showing a cluster wherein the
server gets the list of instances from the db and then iterates
over the list to get the server information. If a shrink
operation is in progress, it can happen that one of the
instances is no longer present when trying to retrieve
the server info, and this causes the show command to throw
a NotFound error.

This is now trapped and the 'missing' server excluded from the list.

Change-Id: I54edc4acac09ca2278f525c08ad0d87576f0549e
Closee-Bug: 1643002