Updated openstack/openstack Project: openstack-infra/project-config  fe21f63cc2e26b7821fca828168b6636931561ff
Have run-tox-with-oslo-master not upgrade to many things

Instead of having the script that gets ran invalidating the
upper constraints, which is currently causing requests to get
upgraded to a newer version that is busting nova (and possible
glance) periodic jobs just respect the constraints (and remove
the oslo libraries from those constraints) and do not try to
run pip with -U so that more than desired gets upgraded.

This seems to localize the packages that get installed and upgraded
to a smaller set (hopefully causing less package variance issues).

Change-Id: Ice723851e5c8861cc1db0dfee2ea665e8727a8a4