Updated openstack/openstack Project: openstack-infra/project-config  843d2d9552172b4c56eb94da56f4aac8e247f0d1
Remove "in-process-" from swift func jobs

Shorten the alternate functional test job names for swift to bring
them under the max shebang length limit for shebangs in the
resulting tox virtualenv in our job workspaces. This will need to be
coordinated with a similar change to swift's tox.ini after the job
configuration is applied.

Change-Id: Icdcf76f7d52e4cce03378c88946c49d28e328547